updated functions getting embeddings to database(embeddings with style transfer added)
