+++ itnt/Constants.py
@@ -9,5 +9,9 @@ TENSORFLOW_EMBED_DISTANCE_NAME = "tf_embed_distances_"
 TENSORFLOW = "tf_embed_"
 ARCFACE_WITH_ST_FOLDER_NAME = "for_arc_with_st"
 TENSORFLOW_WITH_ST_FOLDER_NAME = "for_tf_with_st"
-COUNT_TRAINING_SAMPLES = 78
-COUNT_TESTING_SAMPLES = 32
+COUNT_TRAINING_SAMPLES_TDCS = 78
+COUNT_TESTING_SAMPLES_TDCS = 32
+CUHK_DATASET_NAME = "CUHK"
+TDCS_DATASET_NAME = "TDCS"
+COUNT_TRAINING_SAMPLES_CUHK = 131
+COUNT_TESTING_SAMPLES_CUHK = 57
